refactor(server): extract e2e boot() helper to tests/common/mod.rs
Implements P2-5 of the 2026-05-04 retrospective fix plan, finding
H9. Spinning up the in-process Convergio daemon used to live in 30+
near-identical async fn boot() copies across tests/e2e_*.rs;
adding a field to AppState meant editing every one (the embedder
field was the most recent example, touched 5+ files).

The shared helper now lives at tests/common/mod.rs. Per Cargo
convention each consuming file declares mod common; — no Cargo
metadata change needed. Files keep working byte-for-byte until they
migrate; the 25 unmigrated files have a clean follow-up path.

Five files migrated in this PR (acceptance: 5+):
- e2e_audit.rs, e2e_audit_stream.rs, e2e_quickstart.rs (matched
  signature String, Pool, TempDir; direct call)
- e2e_agent_enrichment.rs (was String, TempDir, Pool; tuple order
  flipped at callsites)
- e2e_bus.rs (was String, TempDir; thin local wrapper drops the
  Pool)

Each file's tests still pass byte-for-byte (3 + 3 + 2 + 6 + 2 = 16
tests across the migrated set). No semantic change; the helper
returns the same AppState shape.

Tracks: ff1225a0-66fd-49a3-8afe-89e9fb922020
